+++ packages/widget/src/CommandMapper.tsx
@@ -21,13 +21,15 @@ import {
   nativeAssetSuperTokenABI,
   superTokenABI,
 } from "./core/index.js";
-import { ChildrenProp, MaxUint256 } from "./utils.js";
+import { MaxUint256 } from "./utils.js";
 import { useWidget } from "./WidgetContext.js";
 
-type CommandMapperProps<TCommand extends Command = Command> = {
+export type CommandMapperProps<TCommand extends Command = Command> = {
   command: TCommand;
-  onMapped?: (contractWrites: ReadonlyArray<ContractWrite>) => void;
-  children?: (contractWrites: ReadonlyArray<ContractWrite>) => ChildrenProp;
+  onMapped?: (result: {
+    commandId: string;
+    contractWrites: ReadonlyArray<ContractWrite>;
+  }) => void;
 };
 
 export function CommandMapper({ command: cmd, ...props }: CommandMapperProps) {
@@ -44,7 +46,6 @@ export function CommandMapper({ command: cmd, ...props }: CommandMapperProps) {
 export function EnableAutoWrapCommandMapper({
   command: cmd,
   onMapped,
-  children,
 }: CommandMapperProps<EnableAutoWrapCommand>) {
   const { getUnderlyingToken } = useWidget();
 
@@ -69,6 +70,7 @@ export function EnableAutoWrapCommandMapper({
         args: [cmd.accountAddress, autoWrapStrategyAddress[cmd.chainId]],
       },
     ],
+    cacheOnBlock: true,
   });
 
   const [wrapScheduleData, allowanceData] = data ?? [];
@@ -122,23 +124,23 @@ export function EnableAutoWrapCommandMapper({
   }, [cmd.id, isSuccess]);
 
   useEffect(
-    () => (isSuccess ? onMapped?.(contractWrites) : void 0),
-    [contractWrites],
+    () =>
+      isSuccess ? onMapped?.({ commandId: cmd.id, contractWrites }) : void 0,
+    [cmd.id, contractWrites, isSuccess],
   );
 
-  return <>{children?.(contractWrites)}</>;
+  return null;
 }
 
 export function WrapIntoSuperTokensCommandMapper({
   command: cmd,
   onMapped,
-  children,
 }: CommandMapperProps<WrapIntoSuperTokensCommand>) {
   const { getSuperToken, getUnderlyingToken } = useWidget();
 
   const isNativeAssetUnderlyingToken = cmd.underlyingToken.isNativeAsset;
 
-  const { data: allowance, isSuccess } = useContractRead(
+  const { data: allowance_, isSuccess } = useContractRead(
     !isNativeAssetUnderlyingToken
       ? {
           chainId: cmd.chainId,
@@ -146,6 +148,7 @@ export function WrapIntoSuperTokensCommandMapper({
           abi: erc20ABI,
           functionName: "allowance",
           args: [cmd.accountAddress, cmd.superTokenAddress],
+          cacheOnBlock: true,
         }
       : undefined,
   );
@@ -168,7 +171,8 @@ export function WrapIntoSuperTokensCommandMapper({
         }),
       );
     } else {
-      if (allowance !== undefined) {
+      if (allowance_ !== undefined) {
+        const allowance = BigInt(allowance_);
         if (allowance < cmd.amountWeiFromUnderlyingTokenDecimals) {
           contractWrites_.push(
             createContractWrite({
@@ -208,25 +212,26 @@ export function WrapIntoSuperTokensCommandMapper({
   }, [cmd.id, isSuccess]);
 
   useEffect(
-    () => (isSuccess ? onMapped?.(contractWrites) : void 0),
-    [contractWrites],
+    () =>
+      isSuccess ? onMapped?.({ commandId: cmd.id, contractWrites }) : void 0,
+    [cmd.id, contractWrites, isSuccess],
   );
 
-  return <>{children?.(contractWrites)}</>;
+  return null;
 }
 
 export function SubscribeCommandMapper({
   command: cmd,
   onMapped,
-  children,
 }: CommandMapperProps<SubscribeCommand>) {
-  const { isSuccess: isSuccessForGetFlowRate, data: existingFlowRate } =
+  const { isSuccess: isSuccessForGetFlowRate, data: existingFlowRate_ } =
     useContractRead({
       chainId: cmd.chainId,
       address: cfAv1ForwarderAddress[cmd.chainId],
       abi: cfAv1ForwarderABI,
       functionName: "getFlowrate",
       args: [cmd.superTokenAddress, cmd.accountAddress, cmd.receiverAddress],
+      cacheOnBlock: true,
     });
 
   const flowRate =
@@ -236,7 +241,9 @@ export function SubscribeCommandMapper({
   const contractWrites = useMemo(() => {
     const contractWrites_: ContractWrite[] = [];
 
-    if (existingFlowRate !== undefined) {
+    if (existingFlowRate_ !== undefined) {
+      const existingFlowRate = BigInt(existingFlowRate_);
+
       if (cmd.transferAmountWei > 0n) {
         contractWrites_.push(
           createContractWrite({
@@ -297,11 +304,14 @@ export function SubscribeCommandMapper({
   }, [cmd.id, isSuccessForGetFlowRate]);
 
   useEffect(
-    () => (isSuccessForGetFlowRate ? onMapped?.(contractWrites) : void 0),
-    [contractWrites],
+    () =>
+      isSuccessForGetFlowRate
+        ? onMapped?.({ commandId: cmd.id, contractWrites })
+        : void 0,
+    [cmd.id, contractWrites, isSuccessForGetFlowRate],
   );
 
-  return <>{children?.(contractWrites)}</>;
+  return null;
 }
 
 const createContractWrite = <